Create Yolo.ipynb
Source code for wrong-way driving has emerged as a significant problem for road safety, thereby demanding robust detection systems. Specifically, this paper provides a solution to wrong-way driving detection by utilizing a deep learning model combining You Only Look Once (YOLOv9) to identify the vehicle and propose a Convolutional Neural Network (CNN) to categorize the behavior. The developed system uses the ability of YOLOv9 to detect objects on the road and then classifies the vehicles using a CNN classifier to determine the direction of travel. The training and validation dataset contains images of moving vehicles in the correct and incorrect directions, as marked by the annotators